Refactor: split create
