+++ cmd/create_data_stream.go
@@ -1,0 +1,18 @@
+// Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
+// or more contributor license agreements. Licensed under the Elastic License;
+// you may not use this file except in compliance with the Elastic License.
+
+package cmd
+
+import "github.com/spf13/cobra"
+
+const createDataStreamLongDescription = `Use this command to create a new data stream.
+
+The command can extend the package with a new data stream using embedded data stream template and wizard.`
+
+func createDataStreamCommandAction(cmd *cobra.Command, args []string) error {
+	cmd.Println("Create a new data stream")
+
+	cmd.Println("Done")
+	return nil
+}
